Changed collision mechanics and it is a little bit better (might revert the change). Also added some ui features and will add GameStates

+++ src/model/Coin.java
@@ -7,10 +7,10 @@ import static ui.GamePanel.TILESIZE;
 public class Coin extends Item {
     Helper helper = new Helper();
 
-    public Coin(int x, int y) {
+    public Coin(float x, float y) {
         image = helper.setup("objects/coin", TILESIZE / 4, TILESIZE / 4);
         this.posX = x;
         this.posY = y;
-        this.hitBox = new Rectangle(x, y, TILESIZE / 4, TILESIZE / 4);
+        this.hitBox = new Rectangle((int) x, (int) y, TILESIZE / 4, TILESIZE / 4);
     }
 }
