+++ src/model/Helper.java
@@ -10,30 +10,30 @@ import static ui.GamePanel.TILESIZE;
 public class Helper {
 
     // Method to draw a BufferedImage
-    public static void draw(Graphics2D g2, BufferedImage img, int posX, int posY) {
+    public static void draw(Graphics2D g2, BufferedImage img, float posX, float posY) {
         Player player = Player.getInstance();
 
         if (isWithinPlayerView(player, posX, posY)) {
-            int screenX = posX - player.getPosX() + player.getScreenX();
-            int screenY = posY - player.getPosY() + player.getScreenY();
-            g2.drawImage(img, screenX, screenY, null);
+            float screenX = posX - player.getPosX() + player.getScreenX();
+            float screenY = posY - player.getPosY() + player.getScreenY();
+            g2.drawImage(img, (int) screenX, (int) screenY, null);
         }
     }
 
     // Method to draw a Rectangle as a filled rectangle
-    public static void draw(Graphics2D g2, Color color, int posX, int posY, int width, int height) {
+    public static void draw(Graphics2D g2, Color color, float posX, float posY, int width, int height) {
         Player player = Player.getInstance();
 
         if (isWithinPlayerView(player, posX, posY)) {
-            int screenX = posX - player.getPosX() + player.getScreenX();
-            int screenY = posY - player.getPosY() + player.getScreenY();
+            float screenX = posX - player.getPosX() + player.getScreenX();
+            float screenY = posY - player.getPosY() + player.getScreenY();
             g2.setColor(color);
-            g2.fillRect(screenX, screenY, width, height);
+            g2.fillRect((int) screenX, (int) screenY, width, height);
         }
     }
 
     // Helper method to check if an object is within the player's view
-    private static boolean isWithinPlayerView(Player player, int posX, int posY) {
+    private static boolean isWithinPlayerView(Player player, float posX, float posY) {
         return posX + 2 * TILESIZE > player.posX - player.getScreenX() &&
                 posX - 2 * TILESIZE < player.posX + player.getScreenX() &&
                 posY + 2 * TILESIZE > player.posY - player.getScreenY() &&
